// src/app/modules/user/user.interface.ts
// Generic type
// type GenericType<T> = Array<T>;

// export type AccountType = {
//     id: string;
//     account_name: string;
// };



export interface IUser {
    name: string;
    email: string;
    password: string;
    mobile: string;
    address: string;
    otp: string;
    token: string;
}